pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["조현일 <himzei@gmail.com>"]
readme = "README.md"
packages = [{include = "airbnb_clone_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"


[tool.poetry.group.dev.dependencies]
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
